+++ src/groups/metadata.rs
@@ -16,36 +16,36 @@ pub struct GroupConfigMetadata {
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupConfigurationMetadata {
-    pub name_max_length: usize,
-    pub description_max_length: usize,
-    pub icon_max_file_size_mb: usize,
-    pub cost: usize,
+    pub name_max_length: u32,
+    pub description_max_length: u32,
+    pub icon_max_file_size_mb: u32,
+    pub cost: u32,
     pub is_using_two_step_webview_component: bool,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRecurringPayoutsConfigurationMetadata {
-    pub max_payout_partners: usize,
+    pub max_payout_partners: u32,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRoleConfigurationMetadata {
-    pub name_max_length: usize,
-    pub description_max_length: usize,
-    pub limit: usize,
-    pub cost: usize,
-    pub min_rank: usize,
-    pub max_rank: usize,
+    pub name_max_length: u32,
+    pub description_max_length: u32,
+    pub limit: u32,
+    pub cost: u32,
+    pub min_rank: u32,
+    pub max_rank: u32,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupNameChangeConfigurationMetadata {
-    pub cost: usize,
-    pub cooldown_in_days: usize,
-    pub ownership_cooldown_in_days: usize,
+    pub cost: u32,
+    pub cooldown_in_days: u32,
+    pub ownership_cooldown_in_days: u32,
 }
 
 /// Gets the metadata for group configuration
@@ -58,12 +58,12 @@ pub async fn config_metadata(jar: &RequestJar) -> Result<GroupConfigMetadata, Bo
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupMetadata {
-    pub group_limit: usize,
-    pub current_group_count: usize,
-    pub group_status_max_length: usize,
-    pub group_post_max_length: usize,
+    pub group_limit: u32,
+    pub current_group_count: u32,
+    pub group_status_max_length: u32,
+    pub group_post_max_length: u32,
     pub is_group_wall_notifications_enabled: bool,
-    pub group_wall_notifications_subscribe_interval_in_milliseconds: usize,
+    pub group_wall_notifications_subscribe_interval_in_milliseconds: u32,
     pub are_profile_groups_hidden: bool,
     pub is_group_details_policy_enabled: bool,
     pub show_previous_group_names: bool,
