+++ src/groups/roles.rs
@@ -15,17 +15,17 @@ use super::user_memberships;
 #[derive(Debug, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRole {
-    pub id: usize,
+    pub id: u32,
     pub name: String,
     pub description: Option<String>,
-    pub rank: usize,
-    pub member_count: Option<usize>,
+    pub rank: u32,
+    pub member_count: Option<u32>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GroupRoleResponse {
-    pub group_id: usize,
+    pub group_id: u32,
     pub roles: Vec<GroupRole>,
 }
 
@@ -33,7 +33,7 @@ pub struct GroupRoleResponse {
 ///
 /// # Error codes
 /// - 1: The group is invalid or does not exist.
-pub async fn roles(jar: &RequestJar, group_id: usize) -> Result<Vec<GroupRole>, Box<Error>> {
+pub async fn roles(jar: &RequestJar, group_id: u32) -> Result<Vec<GroupRole>, Box<Error>> {
     let url = format!("https://groups.roblox.com/v1/groups/{}/roles", group_id);
     let response = jar.get_json::<GroupRoleResponse>(&url).await?;
     Ok(response.roles)
@@ -46,8 +46,8 @@ pub async fn roles(jar: &RequestJar, group_id: usize) -> Result<Vec<GroupRole>, 
 /// - 2: The roleset is invalid or does not exist.
 pub async fn users_on_role(
     jar: &RequestJar,
-    group_id: usize,
-    role_id: usize,
+    group_id: u32,
+    role_id: u32,
     limit: PageLimit,
     sort_order: Option<SortOrder>,
 ) -> Result<Vec<MinimalGroupUser>, Box<Error>> {
@@ -71,8 +71,8 @@ pub async fn users_on_role(
 /// - 200: The user is not in the group.
 pub async fn user_role(
     jar: &RequestJar,
-    group_id: usize,
-    user_id: usize,
+    group_id: u32,
+    user_id: u32,
 ) -> Result<GroupRole, Box<Error>> {
     let roles = user_memberships(jar, user_id).await?;
 
@@ -100,7 +100,7 @@ pub async fn user_role(
 /// - 2: Too many ids in request.
 pub async fn roles_by_id(
     jar: &RequestJar,
-    role_ids: Vec<usize>,
+    role_ids: Vec<u32>,
 ) -> Result<Vec<GroupRole>, Box<Error>> {
     let string_ids = role_ids.iter().map(|id| id.to_string()).collect::<Vec<_>>();
     let url = format!(
